Improvement of new Page Buttons
NewPage Button:
If Site is in a Content Namespace, this will be shown in dialog box
NewSubPage Button:
Dialog box is prefilled with sitename and a slash
NewSubPage Button is not shown on Specialpages, non existing pages and
on redirects.
Titlebox shows search results, but whenever user clicks on 'ok', the raw
value in dialog box will be used as redirection. Less clicks necessary.

removed superflous check for specialpages - specialpages have articleID
of zero.

Change-Id: Ica2e1ce59d1d4dc8dbf82bf9a35ad8ff787397f8