Provide a Chameleon specific version of SkinTemplateOutputPageBeforeExec
T254300

SkinTemplateOutputPageBeforeExec is deprecated but BlueSpice uses this
hook to communicate with Chameleon skin. A new hook is provided to allow
this integrate to continue with minimal disruption.

This hook will be present in Chameleon 3.0.1+ Until we are able to
update to this version this hook is added into Calumma

Change-Id: Ief79d2d27a9c2754fc0f5513eb3a000da3ab3473